+++ src/components/Partners.js
@@ -32,28 +32,27 @@ class Partners extends Component {
         const { classes } = this.props;
         return ( 
            <Grid container>
-               
-                    <Grid container direction="row" justify="center">
-                        <Typography variant="subheading" item xs={3}  className={classes.typography}>Media partners</Typography>
-                        <Grid contianer direction="row" justify="center" align="center">
-                            <img item xs={3} src={business} alt="Business Daily Logo"/>
-                            <img item xs={3} src ={premium} alt="Premium Times Logo"/>
-                            <img item xs={3} src={guardian} alt="The guardian Logo"/>
-                            <img item xs={3} src={nation} alt="The nation Logo" />
-                            <img item xs={3} src={star} alt="The Star Logo"/>
-                            <img item xs={3} src={womanng} alt="Woman NG Logo"/>
-                            <img item xs={3} src={dailynation} alt="Daily Nation Logo"/>
-                            <img item xs={3} src={cable} alt="The Cable Logo"/>
-                        </Grid>
-                    </Grid>
+              <Grid container direction="row" justify="center">
+                <Typography variant="subheading" item xs={3}  className={classes.typography}>Media partners</Typography>
+                <Grid contianer direction="row" justify="center" align="center">
+                  <img item xs={3} src={business} alt="Business Daily Logo"/>
+                  <img item xs={3} src ={premium} alt="Premium Times Logo"/>
+                  <img item xs={3} src={guardian} alt="The guardian Logo"/>
+                  <img item xs={3} src={nation} alt="The nation Logo" />
+                  <img item xs={3} src={star} alt="The Star Logo"/>
+                  <img item xs={3} src={womanng} alt="Woman NG Logo"/>
+                  <img item xs={3} src={dailynation} alt="Daily Nation Logo"/>
+                  <img item xs={3} src={cable} alt="The Cable Logo"/>
+                </Grid>
+              </Grid>
                     
-                    <Grid container direction="column" justify="center" style={{paddingTop:'1rem', paddingBottom:'1rem'}}>
-                        <Typography variant="subheading" className={classes.typography}>In partnership with</Typography>
-                        <div contianer direction="row" justify="center" align="center">
-                            <img src={code} item xs={3} alt="Code For Africa Logo"/>
-                            <img src={innovate} item xs={3} alt="innovateAFRICA Logo"/>
-                        </div>
-                    </Grid>
+              <Grid container direction="column" justify="center" style={{paddingTop:'1rem', paddingBottom:'1rem'}}>
+                <Typography variant="subheading" className={classes.typography}>In partnership with</Typography>
+                  <div contianer direction="row" justify="center" align="center">
+                    <img src={code} item xs={3} alt="Code For Africa Logo"/>
+                    <img src={innovate} item xs={3} alt="innovateAFRICA Logo"/>
+                  </div>
+              </Grid>
            </Grid>
          );
     }
